Replaced double with float. Changed constructor for person to use std::move for strings rather than pass by reference. Changed functions in Person to return by reference so const return operator is useful.

+++ fns.h
@@ -13,7 +13,7 @@ std::vector<std::string> readPaymentsFile(std::ifstream& in);
 void printData(std::ostream &out, const std::vector<std::string> &v);
 Person getNameAndAmount(const std::string& information);
 void createPeople(const std::vector<std::string> &information, std::vector<Person> &people);
-void populateMap(const std::vector<Person> &people, std::unordered_map<std::string , std::vector<std::pair<std::string, double>>> &infoMap);
-void printResults(std::ostream& out, const std::vector<std::string> &names, std::unordered_map<std::string , std::vector<std::pair<std::string, double>>> &infoMap);
+void populateMap(const std::vector<Person> &people, std::unordered_map<std::string , std::vector<std::pair<std::string, float>>> &infoMap);
+void printResults(std::ostream& out, const std::vector<std::string> &names, std::unordered_map<std::string , std::vector<std::pair<std::string, float>>> &infoMap);
 
 #endif //FNS_H
